Add a `TypeAdapterFactory` for Guava collections to `extras`.
Note that the `extras` package is not currently released anywhere. If you want
to use this you must build it from source.

I considered putting this into Gson itself, probably using reflection to
preserve the property that Gson has no non-trivial dependencies. But I think it
is safer to require client code to configure
`GuavaCollectionsTypeAdapterFactory` explicitly, at least for now.